* master: (320 commits)
  A small cleanup in runLoadSave function (plus ForceEditor remnants were removed).
  Fix drawing of radar dishes that broke in r8189.
  Correct mitigation for error resulting in ticket:962, this fix should make game continue
  Remove unused code and correct PIE documentation. Number of texture animations and playback rates do not
  Clean up renderStructure() a bit to remove unused or duplicate code. Use OpenGL matrix commands to
  Added support for ATI-specific version of two-sided stencil extension (GL_ATI_separate_stencil) in pie_DrawShadows. Will decrease CPU usage when using ATI cards.
  updated makefiles for the new original mod.
  Change Foldername from Old School to original.
  Add readme.txt and update the images/frontend0.png
  Fix bug	that would allow artillery to use sensors without regard to actual visibility of target.
  Stop stripping debug information from release binaries in Mac OS X.
  Fix ticket:883 assert on savegame load during campaign. The problem is that you load droids
  Fix scrGroupAddAreaNoGroup script function, was called as scrGroupAddArea instead.
  Explicitly link against libintl.
  Display SQLite version during startup. Addresses #936.
  Add the new Old School Mod to the SVN (trunk and branch).
  Pies for Multiple muzzle flashes feature #824 (part 2).
  Additional code changes for Multiple muzzle flashes feature #824.
  Fix bug #939: Cannot compile in Linux.
  Partially fix bug #763: Units not going to specified repair facility.
  ...

Conflicts:
	data/mods/multiplay/aivolution/multiplay/skirmish/player0.slo
	data/mods/multiplay/aivolution/multiplay/skirmish/player1.slo
	data/mods/multiplay/aivolution/multiplay/skirmish/player2.slo
	data/mods/multiplay/aivolution/multiplay/skirmish/player3.slo
	data/mods/multiplay/aivolution/multiplay/skirmish/player4.slo
	data/mods/multiplay/aivolution/multiplay/skirmish/player5.slo
	data/mods/multiplay/aivolution/multiplay/skirmish/player6.slo
	data/mods/multiplay/aivolution/multiplay/skirmish/player7.slo
	src/Makefile.am
	src/objmem.c
	src/scriptai.h
	src/scriptfuncs.c
	src/scripttabs.c
	win32/Warzone2100.sln

+++ lib/ivis_opengl/piematrix.c
@@ -298,7 +298,7 @@ void pie_PerspectiveEnd(void)
 {
 	glMatrixMode(GL_PROJECTION);
 	glLoadIdentity();
-	glOrtho(0, pie_GetVideoBufferWidth(), pie_GetVideoBufferHeight(), 0, 1, -1);
+	glOrtho(0.0f, (double) pie_GetVideoBufferWidth(), (double) pie_GetVideoBufferHeight(), 0.0f, 1.0f, -1.0f);
 	glMatrixMode(GL_MODELVIEW);
 }
